fix C:\projects\metaSMT\src\metaSMT/backend/Z3_Backend.hpp(423): warning C4244: 'argument': conversion from 'uint64_t' to 'unsigned int', possible loss of data [C:\projects\metaSMT\build\tests\direct_Z3.vcxproj]

+++ src/metaSMT/backend/Z3_Backend.hpp
@@ -416,21 +416,21 @@ namespace metaSMT {
       }
 
       result_type operator() (bvtags::extract_tag const &,
-                              uint64_t upper,
-                              uint64_t lower,
+                              unsigned int upper,
+                              unsigned int lower,
                               result_type const &e) {
         return z3::to_expr(ctx_,
           Z3_mk_extract(ctx_, upper, lower, z3::expr(e)));
       }
 
       result_type operator() (bvtags::zero_extend_tag const &,
-                              uint64_t width,
+                              unsigned int width,
                               result_type e) {
         return z3::to_expr(ctx_, Z3_mk_zero_ext(ctx_, width, z3::expr(e)));
       }
 
       result_type operator() (bvtags::sign_extend_tag const &, 
-                              uint64_t width,
+                              unsigned int width,
                               result_type e) {
         return z3::to_expr(ctx_, Z3_mk_sign_ext(ctx_, width, z3::expr(e)));
       }
